Added some sprites and rebinded input actions

+++ Assets/Scripts/Core/ECS/Components/Events/Units/DamageEventComponent.cs
@@ -4,7 +4,7 @@ namespace Core.ECS.Events
 {
     internal struct DamageEventComponent
     {
-        public float Damage;
+        public byte Damage;
         public GameObject Target;
         public GameObject Source;
     }
